tests: Don't sleep in KillIntegrationTest
There was a `sleep(100)` to wait for the query which should be killed to
start executing.
Instead of that it will now always wait for a `sys.jobs` entry to
appear.
That was already the case anyway for tests using `KILL <jobId>` instead
of `KILL ALLL`.

+++ sql/src/test/java/io/crate/integrationtests/KillIntegrationTest.java
@@ -21,12 +21,15 @@
 
 package io.crate.integrationtests;
 
+import com.google.common.util.concurrent.SettableFuture;
 import io.crate.action.sql.SQLActionException;
+import io.crate.action.sql.SQLRequest;
 import io.crate.action.sql.SQLResponse;
 import io.crate.exceptions.Exceptions;
 import io.crate.plugin.SQLPlugin;
 import io.crate.testing.UseJdbc;
 import io.crate.testing.plugin.CrateTestingPlugin;
+import org.elasticsearch.action.ActionFuture;
 import org.elasticsearch.plugins.Plugin;
 import org.elasticsearch.threadpool.ThreadPool;
 import org.junit.After;
@@ -38,10 +41,7 @@ import org.junit.rules.TemporaryFolder;
 import javax.annotation.Nullable;
 import java.util.Collection;
 import java.util.UUID;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicReference;
 
 import static com.carrotsearch.randomizedtesting.RandomizedTest.$;
 import static org.hamcrest.Matchers.*;
@@ -85,46 +85,22 @@ public class KillIntegrationTest extends SQLTransportIntegrationTest {
     }
 
     private void assertGotCancelled(final String statement, @Nullable final Object[] params, boolean killAll) throws Exception {
-        final ExecutorService executor = Executors.newSingleThreadExecutor();
-        final AtomicReference<Throwable> thrown = new AtomicReference<>();
         try {
-            executor.submit(new Runnable() {
-                @Override
-                public void run() {
-                    try {
-                        execute(statement, params);
-                    } catch (Throwable e) {
-                        Throwable unwrapped = Exceptions.unwrap(e);
-                        thrown.compareAndSet(null, unwrapped);
-                    }
-                }
-            });
-            Thread.sleep(100L);
+            ActionFuture<SQLResponse> future = sqlExecutor.execute(new SQLRequest(statement, params));
+            String jobId = waitForJobEntry(statement);
+            if (jobId == null) {
+                // query finished too fast
+                return;
+            }
             if (killAll) {
                 execute("kill all");
             } else {
-                assertBusy(new Runnable() {
-                    @Override
-                    public void run() {
-                        SQLResponse logResponse = execute("select * from sys.jobs where stmt = ?", $(statement));
-                        if (logResponse.rowCount() == 0) {
-                            logResponse = execute("select * from sys.jobs_log where stmt = ?", $(statement));
-                            if (logResponse.rowCount() > 0L) {
-                                // query finished before jobId could be retrieved
-                                // finishing without killing - test will pass which is okay because it is not deterministic by design
-                                return;
-                            }
-                        }
-                        assertThat(logResponse.rowCount(), greaterThan(0L));
-                        String jobId = logResponse.rows()[0][0].toString();
-                        execute("kill ?", new Object[]{jobId});
-                    }
-                });
+                execute("kill ?", $(jobId));
             }
-            executor.shutdown();
-            executor.awaitTermination(5L, TimeUnit.SECONDS);
-            Throwable exception = thrown.get();
-            if (exception != null) {
+            try {
+                future.get(10, TimeUnit.SECONDS);
+            } catch (Throwable exception) {
+                exception = Exceptions.unwrap(exception); // wrapped in ExecutionException
                 assertThat(exception, instanceOf(SQLActionException.class));
                 assertThat(exception.toString(), anyOf(
                         containsString("Job killed"), // CancellationException
@@ -133,9 +109,36 @@ public class KillIntegrationTest extends SQLTransportIntegrationTest {
                 ));
             }
         } finally {
-            executor.shutdownNow();
             waitUntilThreadPoolTasksFinished(ThreadPool.Names.SEARCH);
         }
+    }
+
+    /**
+     * Wait for a statement to appear in sys.jobs.
+     * If the query finished execution (is in sys.jobs_log) it will return null
+     */
+    @Nullable
+    private String waitForJobEntry(final String statement) throws Exception {
+        final SettableFuture<String> jobIdFuture = SettableFuture.create();
+        assertBusy(new Runnable() {
+            @Override
+            public void run() {
+                SQLResponse logResponse = execute("select * from sys.jobs where stmt = ?", $(statement));
+                if (logResponse.rowCount() == 0) {
+                    logResponse = execute("select * from sys.jobs_log where stmt = ?", $(statement));
+                    if (logResponse.rowCount() > 0L) {
+                        // query finished before jobId could be retrieved
+                        // finishing without killing - test will pass which is okay because it is not deterministic by design
+                        jobIdFuture.set(null);
+                        return;
+                    }
+                }
+                assertThat(logResponse.rowCount(), greaterThan(0L));
+                String jobId = logResponse.rows()[0][0].toString();
+                jobIdFuture.set(jobId);
+            }
+        });
+        return jobIdFuture.get(10, TimeUnit.SECONDS);
     }
 
     @Test
